Refactor composer into package and update docs
