+++ composer/__init__.py
@@ -1,0 +1,16 @@
+from .io import load_image, save_image
+from .ml import generate_depth, depth_to_normal, generate_alpha, decompose_albedo_shading
+from .pipeline import process_image
+from .cli import build_arg_parser, main
+
+__all__ = [
+    "load_image",
+    "save_image",
+    "generate_depth",
+    "depth_to_normal",
+    "generate_alpha",
+    "decompose_albedo_shading",
+    "process_image",
+    "build_arg_parser",
+    "main",
+]
