+++ input/README.md
@@ -1,0 +1,1 @@
+Place images to process here.
