+++ output/README.md
@@ -1,0 +1,1 @@
+Generated outputs will appear here.
