+++ requirements.txt
@@ -1,8 +1,9 @@
-torch
-torchvision
-opencv-python
-numpy
-pillow
-midas
+torch==2.2.1
+torchvision==0.17.1
+opencv-python==4.12.0.88
+numpy==2.2.6
+pillow==11.3.0
+midas==1.1.2
+# Optional U2Net model (may require manual install)
 u2net
 # iiwnet (optional)
